feat: move CLI implementation into package
Move the CLI code from the repository root into gemini_webapi.cli so it is

importable after installation. Keep cli.py as a thin wrapper that adds

src/ to sys.path for source checkouts.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
